# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/coveredpersons": {
            "get": {
                "description": "list coveredperson entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List coveredperson entities",
                "operationId": "list-coveredperson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.CoveredPerson"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create coveredperson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create coveredperson",
                "operationId": "create-coveredperson",
                "parameters": [
                    {
                        "description": "CoveredPerson entity",
                        "name": "coveredperson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CoveredPerson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.CoveredPerson"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/coveredpersons/{id}": {
            "get": {
                "description": "get coveredperson by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a coveredperson entity by ID",
                "operationId": "get-coveredperson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CoveredPerson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.CoveredPerson"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update coveredperson by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a coveredperson entity by ID",
                "operationId": "update-coveredperson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CoveredPerson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CoveredPerson entity",
                        "name": "coveredperson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.CoveredPerson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.CoveredPerson"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get coveredperson by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a coveredperson entity by ID",
                "operationId": "delete-coveredperson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CoveredPerson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.CoveredPerson": {
            "type": "object",
            "properties": {
                "certificate": {
                    "type": "integer"
                },
                "fund": {
                    "type": "integer"
                },
                "patient": {
                    "type": "integer"
                },
                "schemeType": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Bloodtype": {
            "type": "object",
            "properties": {
                "BTname": {
                    "description": "BTname holds the value of the \"BTname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BloodtypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BloodtypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BloodtypeEdges": {
            "type": "object",
            "properties": {
                "frombloodtype": {
                    "description": "Frombloodtype holds the value of the frombloodtype edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patient"
                    }
                }
            }
        },
        "ent.Certificate": {
            "type": "object",
            "properties": {
                "Certificate_Name": {
                    "description": "CertificateName holds the value of the \"Certificate_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CertificateQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CertificateEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.CertificateEdges": {
            "type": "object",
            "properties": {
                "certificateCoveredPerson": {
                    "description": "CertificateCoveredPerson holds the value of the Certificate_CoveredPerson edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CoveredPerson"
                    }
                }
            }
        },
        "ent.CoveredPerson": {
            "type": "object",
            "properties": {
                "certificate_id": {
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CoveredPersonQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CoveredPersonEdges"
                },
                "fund_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "patient_id": {
                    "type": "integer"
                },
                "schemeType_id": {
                    "type": "integer"
                }
            }
        },
        "ent.CoveredPersonEdges": {
            "type": "object",
            "properties": {
                "certificate": {
                    "description": "Certificate holds the value of the Certificate edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Certificate"
                },
                "fund": {
                    "description": "Fund holds the value of the Fund edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Fund"
                },
                "patient": {
                    "description": "Patient holds the value of the Patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "schemeType": {
                    "description": "SchemeType holds the value of the SchemeType edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SchemeType"
                }
            }
        },
        "ent.Department": {
            "type": "object",
            "properties": {
                "Department_Name": {
                    "description": "DepartmentName holds the value of the \"Department_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DepartmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DepartmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Diagnose": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DiagnoseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DiagnoseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DiagnoseEdges": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the department edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "disease": {
                    "description": "Disease holds the value of the disease edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Disease"
                },
                "doctor": {
                    "description": "Doctor holds the value of the doctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                }
            }
        },
...
        "ent.Fund": {
            "type": "object",
            "properties": {
                "Fund_Name": {
                    "description": "FundName holds the value of the \"Fund_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the FundQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.FundEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.FundEdges": {
            "type": "object",
            "properties": {
                "fundCoveredPerson": {
                    "description": "FundCoveredPerson holds the value of the Fund_CoveredPerson edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CoveredPerson"
                    }
                }
            }
        },
        "ent.Gender": {
            "type": "object",
            "properties": {
                "Gname": {
                    "description": "Gname holds the value of the \"Gname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GenderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GenderEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.GenderEdges": {
            "type": "object",
            "properties": {
                "fromgender": {
                    "description": "Fromgender holds the value of the fromgender edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patient"
                    }
                }
            }
        },
...
        "ent.Nurse": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NurseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NurseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "nurse_email": {
                    "description": "NurseEmail holds the value of the \"nurse_email\" field.",
                    "type": "string"
                },
                "nurse_name": {
                    "description": "NurseName holds the value of the \"nurse_name\" field.",
                    "type": "string"
                },
                "nurse_password": {
                    "description": "NursePassword holds the value of the \"nurse_password\" field.",
                    "type": "string"
                },
                "nurse_tel": {
                    "description": "NurseTel holds the value of the \"nurse_tel\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Patient": {
            "type": "object",
            "properties": {
                "Patient_age": {
                    "description": "PatientAge holds the value of the \"Patient_age\" field.",
                    "type": "integer"
                },
                "Patient_height": {
                    "description": "PatientHeight holds the value of the \"Patient_height\" field.",
                    "type": "number"
                },
                "Patient_name": {
                    "description": "PatientName holds the value of the \"Patient_name\" field.",
                    "type": "string"
                },
                "Patient_weight": {
                    "description": "PatientWeight holds the value of the \"Patient_weight\" field.",
                    "type": "number"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PatientEdges": {
            "type": "object",
            "properties": {
                "bloodtype": {
                    "description": "Bloodtype holds the value of the bloodtype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bloodtype"
                },
                "frompatient": {
                    "description": "Frompatient holds the value of the frompatient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Rent"
                },
                "gender": {
                    "description": "Gender holds the value of the gender edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "patientCoveredPerson": {
                    "description": "PatientCoveredPerson holds the value of the Patient_CoveredPerson edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CoveredPerson"
                    }
                },
                "patientDiagnose": {
                    "description": "PatientDiagnose holds the value of the patient_diagnose edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Diagnose"
                    }
                },
                "prefix": {
                    "description": "Prefix holds the value of the prefix edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prefix"
                }
            }
        },
        "ent.Prefix": {
            "type": "object",
            "properties": {
                "Pname": {
                    "description": "Pname holds the value of the \"Pname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrefixQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrefixEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PrefixEdges": {
            "type": "object",
            "properties": {
                "fromprefix": {
                    "description": "Fromprefix holds the value of the fromprefix edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patient"
                    }
                }
            }
        },
        "ent.Rent": {
            "type": "object",
            "properties": {
                "added_time": {
                    "description": "AddedTime holds the value of the \"added_time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "patient_id": {
                    "type": "integer"
                }
            }
        },
        "ent.RentEdges": {
            "type": "object",
            "properties": {
                "nurse": {
                    "description": "Nurse holds the value of the nurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "room": {
                    "description": "Room holds the value of the room edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Room"
                }
            }
        },
        "ent.Room": {
            "type": "object",
            "properties": {
                "building": {
                    "description": "Building holds the value of the \"building\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomEdges"
                },
                "floor": {
                    "description": "Floor holds the value of the \"floor\" field.",
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.SchemeType": {
            "type": "object",
            "properties": {
                "SchemeType_Name": {
                    "description": "SchemeTypeName holds the value of the \"SchemeType_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SchemeTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SchemeTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.SchemeTypeEdges": {
            "type": "object",
            "properties": {
                "schemeTypeCoveredPerson": {
                    "description": "SchemeTypeCoveredPerson holds the value of the SchemeType_CoveredPerson edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CoveredPerson"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}